+++ src/plugins/trigger/trigger.c
@@ -700,11 +700,11 @@ trigger_regex_split (const char *str_regex,
 
         /* search the delimiter (which can be more than one char) */
         pos = weechat_utf8_next_char (ptr_regex);
-        while (pos[0] && (weechat_string_charcmp (ptr_regex, pos) == 0))
+        while (pos && pos[0] && (weechat_string_charcmp (ptr_regex, pos) == 0))
         {
             pos = weechat_utf8_next_char (pos);
         }
-        if (!pos[0])
+        if (!pos || !pos[0])
             goto format_error;
         delimiter = weechat_strndup (ptr_regex, pos - ptr_regex);
         if (!delimiter)